More code review fixes, move arguments in each class.
