+++ READMEs/main-flow.md
@@ -71,7 +71,7 @@ exchange = datatoken.create_exchange(price, OCEAN.address, {"from": alice})
 datatoken.mint(alice, to_wei(100), {"from": alice_wallet})
 datatoken.approve(exchange.address, to_wei(100), {"from": alice})
 
-# D.3 Bob lets exchange pull the OCEAN needed 
+# D.3 Bob lets exchange pull the OCEAN needed
 OCEAN_needed = exchange.BT_needed(to_wei(1), consume_market_fee=0)
 OCEAN.approve(exchange.address, OCEAN_needed, {"from":bob})
 
@@ -139,7 +139,7 @@ datatoken = Datatoken(config, datatoken_address)
 Data NFTs implement ERC721 functionality, and ERC725 which extends it.
 
 ERC721:
-- Basic spec of a non-fungible token (NFT) 
+- Basic spec of a non-fungible token (NFT)
 - Official spec is at [erc721.org](https://erc721.org/)
 - Solidity interface is in Ocean contracts repo as [IERC721Template.sol](https://github.com/oceanprotocol/contracts/blob/main/contracts/interfaces/IERC721Template.sol)
 
@@ -211,7 +211,7 @@ url_file = UrlFile(
 )
 
 # Publish data asset
-from ocean_lib.ocean.ocean_assets import DatatokenArguments
+from ocean_lib.models.datatoken import DatatokenArguments
 _, _, ddo = ocean.assets.create(
     metadata,
     alice,
@@ -234,7 +234,7 @@ You can control this during create():
 
 Calling `create()` like above generates a data NFT, a datatoken for that NFT, and a ddo. This is the most common case. However, sometimes you may want _just_ the data NFT, e.g. if using a data NFT as a simple key-value store. Here's how:
 ```python
-from ocean_lib.models.arguments import DataNFTArguments
+from ocean_lib.models.data_nft import DataNFTArguments
 data_nft = ocean.data_nft_factory.create(DataNFTArguments('NFT1', 'NFT1'), alice)
 ```
 
@@ -245,7 +245,7 @@ If you call `create()` after this, you can pass in an argument `data_nft_address
 Calling `create()` like above generates a data NFT, a datatoken for that NFT, and a ddo object. However, we may want a second datatoken. Or, we may have started with _just_ the data NFT, and want to add a datatoken to it. Here's how:
 
 ```python
-from ocean_lib.models.arguments import DatatokenArguments
+from ocean_lib.models.datatoken import DatatokenArguments
 datatoken = data_nft.create_datatoken(DatatokenArguments("Datatoken 1", "DT1"), alice)
 ```
 
@@ -310,7 +310,7 @@ print(exchange.fees_info)
 It will output something like:
 ```text
 >>> print(exchange.details)
-ExchangeDetails: 
+ExchangeDetails:
   datatoken = 0xdA3cf7aE9b28E1A9B5F295201d9AcbEf14c43019
   base_token = 0x24f42342C7C171a66f2B7feB5c712471bED92A97
   fixed_rate (price) = 1.0 (1000000000000000000 wei)
@@ -325,7 +325,7 @@ ExchangeDetails:
   owner = 0x02354A1F160A3fd7ac8b02ee91F04104440B28E7
 
 >>> print(exchange.fees_info)
-FeesInfo: 
+FeesInfo:
   publish_market_fee = 0.0 (0 wei)
   publish_market_fee_available = 0.0 (0 wei)
   publish_market_fee_collector = 0x02354A1F160A3fd7ac8b02ee91F04104440B28E7
