+++ ocean_lib/models/test/test_datatoken_order_both_templates.py
@@ -6,8 +6,7 @@ from datetime import datetime
 
 import pytest
 
-from ocean_lib.models.arguments import FeeTokenInfo
-from ocean_lib.models.datatoken import Datatoken
+from ocean_lib.models.datatoken import Datatoken, FeeTokenInfo
 from ocean_lib.ocean.util import from_wei, get_address_of_type, to_wei
 from ocean_lib.web3_internal.constants import MAX_UINT256
 from tests.resources.helper_functions import deploy_erc721_erc20, get_mock_provider_fees
