+++ ocean_lib/ocean/test/test_ocean_assets.py
@@ -17,9 +17,9 @@ from ocean_lib.agreements.service_types import ServiceTypes
 from ocean_lib.assets.ddo import DDO
 from ocean_lib.example_config import DEFAULT_PROVIDER_URL
 from ocean_lib.exceptions import AquariusError, InsufficientBalance
-from ocean_lib.models.arguments import DataNFTArguments, FeeTokenInfo
+from ocean_lib.models.data_nft import DataNFTArguments
 from ocean_lib.models.data_nft_factory import DataNFTFactoryContract
-from ocean_lib.ocean.ocean_assets import DatatokenArguments
+from ocean_lib.models.datatoken import DatatokenArguments, FeeTokenInfo
 from ocean_lib.ocean.util import get_address_of_type
 from ocean_lib.services.service import Service
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
